Renamed test dirs, changed import to match platform-common dirname change
