+++ sonic-storagemond/tests/mocked_libs/sonic_platform_base/sonic_storage/storage_base.py
